04_shipping_traffic_anomaly_detection/requirements.txt
matplotlib
scipy
pandas
sklearn
keras
tensorflow
pyephem
jinja2
Cython
numpy
GEOS
shapely
pyshp
six
PROJ
psycopg2-binary
PyQt5
pydot
